+++ CHANGELOG.md
@@ -4,6 +4,26 @@ If you like this project and find it useful, please consider giving it a star on
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2024-11-21
+
+### Added
+
+- [edge]: Update to support Matterbridge edge (matter.js new API).
+- [matter]: Added bridgedNode and powerSource device types to the cover.
+
+### Changed
+
+- [package]: Requires matterbridge 1.6.0.
+- [package]: Updated dependencies.
+
+### Fixed
+
+- [somfy]: Fixed stop sent when the target is fully open or fully closed.
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="./yellow-button.png" alt="Buy me a coffee" width="120">
+</a>
+
 ## [1.1.1] - 2024-10-12
 
 ### Fixed